requirements.txt
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
idna==2.10
pycodestyle==2.6.0
pycryptodome==3.9.8
requests==2.24.0
toml==0.10.1
urllib3==1.25.10
